move addr reserve/rlease to a10_network_task
