+++ a10_octavia/tests/unit/controller/worker/tasks/test_server_tasks.py
@@ -14,7 +14,6 @@
 
 
 import imp
-import sys
 
 try:
     from unittest import mock
@@ -45,7 +44,6 @@ MEMBER = o_data_models.Member(
 KEY_ARGS = {'server': utils.meta(MEMBER, 'server', {'conn_resume': None, 'conn_limit': 64000000})}
 SERVER_NAME = '{}_{}'.format(MEMBER.project_id[:5],
                              MEMBER.ip_address.replace('.', '_'))
-NAT_POOL = data_models.NATPool(port_id=a10constants.MOCK_PORT_ID)
 
 
 class TestHandlerServerTasks(base.BaseTaskTestCase):
@@ -209,32 +207,3 @@ class TestHandlerServerTasks(base.BaseTaskTestCase):
             POOL.id, SERVER_NAME, MEMBER.protocol_port)
         self.client_mock.slb.server.port.delete.assert_called_with(
             SERVER_NAME, MEMBER.protocol_port, pool_protocol_tcp)
-
-    def test_reserver_subnet_addr(self):
-        mock1 = task.MemberReserveSubnetAddr()
-        if sys.version_info[0] < 3:
-            mock1._network_driver = self.client_mock
-        else:
-            mock1.network_driver = self.client_mock
-        flavor = {"pool_name": "p1", "start_address": "1.1.1.1", "end_address": "1.1.1.2"}
-        mock1.execute(MEMBER, flavor)
-        self.client_mock.reserve_subnet_addresses.assert_called_with(
-            MEMBER.subnet_id, ["1.1.1.1", "1.1.1.2"])
-
-    def test_release_subnet_addr_referenced(self):
-        mock = task.MemberReleaseSubnetAddr()
-        flavor = {"pool_name": "p1", "start_address": "1.1.1.1", "end_address": "1.1.1.2"}
-        NAT_POOL.member_ref_count = 2
-        ret_val = mock.execute(MEMBER, flavor, NAT_POOL)
-        self.assertEqual(ret_val, None)
-
-    def test_release_subnet_addr(self):
-        mock = task.MemberReleaseSubnetAddr()
-        if sys.version_info[0] < 3:
-            mock._network_driver = self.client_mock
-        else:
-            mock.network_driver = self.client_mock
-        flavor = {"pool_name": "p1", "start_address": "1.1.1.1", "end_address": "1.1.1.2"}
-        NAT_POOL.member_ref_count = 1
-        mock.execute(MEMBER, flavor, NAT_POOL)
-        self.client_mock.delete_port.assert_called_with(NAT_POOL.port_id)
